refactor: calcを移行
